+++ controllers/schools.js
@@ -1,5 +1,6 @@
 const schoolsRouter = require("express").Router();
 const School = require("../models/school");
+const auth = require("../utils/auth");
 
 schoolsRouter.get("/", async (req, res, next) => {
   try {
@@ -12,6 +13,12 @@ schoolsRouter.get("/", async (req, res, next) => {
 
 schoolsRouter.post("/", async (req, res, next) => {
   try {
+    const token = auth.getTokenFrom(req);
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
     const addedSchool = new School({
       name: req.body.name,
       link: req.body.link,
@@ -27,6 +34,12 @@ schoolsRouter.post("/", async (req, res, next) => {
 
 schoolsRouter.put("/:id", async (req, res, next) => {
   try {
+    const token = auth.getTokenFrom(req);
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
     const school = {
       name: req.body.name,
       link: req.body.link,
@@ -48,6 +61,12 @@ schoolsRouter.put("/:id", async (req, res, next) => {
 
 schoolsRouter.delete("/:id", async (req, res, next) => {
   try {
+    const token = auth.getTokenFrom(req);
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
     await School.findByIdAndRemove(req.params.id);
     res.status(204).end();
   } catch (error) {
